Adding controller input
Adding controller input part one. The controller is loaded, input is detected, input actually activates stuff, however the behavior from the input isn't what is desired, need some time to figure it out - should be incredibly simple but I may be too lazy to work on it rn and just call it good because the input loading is what I was more worried about.