First basic implementation if tile map and player
Created a basic first implementation of the tile map and the player. This is exactly like the one I created before from Handmade Hero. However, from here on out is where we will start to deviate as the tilemaps are not going to be randomly generated and instead static and hand crafted.