Tilemap edits
Continuing to develop the tilemap and such. The camera now focuses in the middle, the number of tiles per chunk is correct so that it fills the screen. The current issue is the camera which seems to center in the x and y when the player reaches the end of the screen, though I haven't fully tested that so I'm not sure if that's true. But I do believe there is still something wrong with it from the previous testing.