Tilemap Creation
Submitting before I redo a bunch of stuff pertaining to the creation of the tile map. It is currently a sparse storage system as according to Handmade Hero, I realized this is not what I really need and is making it more difficult to store the tile map in a file since the idea was that it was randomly generated, and randomly generated is not what I want.