Animations added
Much to my shagrin, I added in animations in the easy way which will make implementing an atlas for all of our tiles much more difficult, but oh well we'll cross that bridge when we come to it. Probably still need to fiddle around with the timings of everything in the animations will leave that for another submit.